refactor: unify sidebar and header
